added spiritual communication category
