+++ src/components/ClassesPage/ClassDetails.jsx
@@ -3,6 +3,7 @@ import { useParams } from 'react-router-dom';
 import { classesData } from '../constants';
 import textsEng from '../../texts/strings_eng';
 import DetailsView from '../shared/DetailsViewPage';
+import SpiritualCommunication from '../categories/SpiritualCommunication';
 
 const ClassDetails = () => {
   const { classId } = useParams();
@@ -10,7 +11,14 @@ const ClassDetails = () => {
     (classItem) => classItem.id === parseInt(classId));
 
   return (
-    <DetailsView data={classItem} detailsViewBtnText={textsEng['classesPage.classDetails.bookNow.button']} />
+    <div className="class-details-view-container">
+      <DetailsView
+        data={classItem}
+        detailsViewBtnText={textsEng['classesPage.classDetails.bookNow.button']}
+      />
+      {classItem.category && classItem.category === 'Spiritual Communication' &&
+        <SpiritualCommunication />}
+    </div>
   );
 };
