+++ src/components/categories/IchingOracleCards/IchingOracleCards.css
@@ -1,0 +1,3 @@
+.iching-cards-info-header {
+  margin-top: 80px;
+}
