+++ src/components/categories/IchingOracleCards/index.jsx
@@ -1,0 +1,37 @@
+import React from 'react';
+import ReactPlayer from 'react-player';
+import YouTube from 'react-youtube';
+import textsEng from '../../../texts/strings_eng';
+import './IchingOracleCards.css';
+
+const IchingOracleCardsInfo = () => {
+  const options = {
+    height: '390',
+    width: '640',
+    playerVars: {
+      autoplay: 1,
+      controls: 1,
+    },
+  };
+
+  const _onReady = (event) => {
+    event.target.pauseVideo();
+  };
+
+  return (
+    <div className="iching-cards-info-section">
+      <h2 className="iching-cards-info-header">{textsEng['products.category.iChingCards.header']}</h2>
+      <YouTube videoId="tdYz0zbgmT8" options={options} onReady={_onReady}/>
+      <p>{textsEng['products.iChingCardsInstructions.about.content.p1']}</p>
+      <p>{textsEng['products.iChingCardsInstructions.about.content.p2']}</p>
+      <p>{textsEng['products.iChingCardsInstructions.about.content.p3']}</p>
+      <p>{textsEng['products.iChingCardsInstructions.about.content.p4']}</p>
+      <ReactPlayer url={process.env.PUBLIC_URL + '/I Ching 1 mins video.MP4'} controls={true} />
+      <h2 className="iching-cards-info-header">{textsEng['products.iChingCardsInstructions.howTo.header']}</h2>
+      <YouTube videoId="F8LjnvWI0vI" options={options} onReady={_onReady}/>
+      <YouTube videoId="AmU1eq20SgY" options={options} onReady={_onReady}/>
+    </div>
+  );
+};
+
+export default IchingOracleCardsInfo;
