+++ src/components/categories/SingingBowls/SingingBowls.css
@@ -1,0 +1,5 @@
+.singing-bowls-section {
+  width: 70%;
+  margin: 0 auto;
+  padding: 0 50px;
+}
