+++ src/components/categories/SpiritualCommunication/SpiritualCommunication.css
@@ -1,0 +1,5 @@
+.spiritual-communication-section {
+  width: 70%;
+  margin: 0 auto;
+  padding: 0 50px;
+}
