+++ src/components/categories/SubconsciousIntegration/SubconsciousIntegration.css
@@ -1,0 +1,9 @@
+.subconscious-integration-section {
+  width: 70%;
+  margin: 0 auto;
+  padding: 0 50px;
+}
+
+.subconscious-integration-sub-section:last-child {
+  margin-top: 50px;
+}
