--- src/components/products/ProductsPage/category/IchingOracleCards/IchingOracleCards.css
@@ -1,3 +1,0 @@
-.iching-cards-info-header {
-  margin-top: 80px;
-}
