--- src/components/products/ProductsPage/category/SingingBowls/SingingBowls.css
@@ -1,5 +1,0 @@
-.singing-bowls-section {
-  width: 70%;
-  margin: 0 auto;
-  padding: 0 50px;
-}
